+++ sites/partners/__tests__/DatesFormatter.test.ts
@@ -27,8 +27,8 @@ describe("DatesFormatter", () => {
       applicationDueTimeField: dueTime,
     }
     const applicationDueDate = formatData(data).applicationDueDate
-    expect(applicationDueDate).toEqual(createDate(dueDate))
-    expect(formatData(data).applicationDueTime).toEqual(createTime(applicationDueDate, dueTime))
+    expect(applicationDueDate).toEqual(createTime(applicationDueDate, dueTime))
+    expect(formatData(data).applicationDueDate).toEqual(createTime(applicationDueDate, dueTime))
   })
 
   it("should format postmarkedApplicationsReceivedByDate", () => {