+++ sites/partners/src/listings/PaperListingDetails/sections/DetailAdditionalEligibility.tsx
@@ -77,7 +77,7 @@ const DetailAdditionalEligibility = () => {
             ) : (
               <MinimalTable
                 id="buildingSelectionCriteriaTable"
-                headers={{ url: t("t.url") }}
+                headers={{ url: "t.url" }}
                 data={[
                   {
                     url: listing.buildingSelectionCriteria,