+++ sites/public/lib/ApplicationConductor.ts
@@ -14,6 +14,19 @@ import SelectedPreferencesStep from "./SelectedPreferencesStep"
 import PreferencesAllStep from "./PreferencesAllStep"
 import ProgramsStep from "./ProgramsStep"
 
+export const loadApplicationFromAutosave = () => {
+  if (typeof window != "undefined") {
+    const autosavedApplication = window.sessionStorage.getItem("bloom-app-autosave")
+    if (autosavedApplication) {
+      const application = JSON.parse(autosavedApplication)
+      application.loaded = true
+      return application
+    }
+  }
+
+  return null
+}
+
 export const loadSavedListing = () => {
   if (typeof window != "undefined") {
     const savedListing = window.sessionStorage.getItem("bloom-app-listing")
@@ -181,6 +194,7 @@ export default class ApplicationConductor {
     // NOTE: had to remove timeout because of Next doing full-page reloads in
     // some cases. Need to revisit after upgrading to v10
     if (typeof window != "undefined") {
+      window.sessionStorage.setItem("bloom-app-autosave", JSON.stringify(this.application))
       if (this.listing) {
         window.sessionStorage.setItem("bloom-app-listing", JSON.stringify(this.listing))
       }
@@ -192,6 +206,7 @@ export default class ApplicationConductor {
     this.listing = {} as Listing
     this.currentStepIndex = 0
     if (typeof window != "undefined") {
+      window.sessionStorage.removeItem("bloom-app-autosave")
       window.sessionStorage.removeItem("bloom-app-listing")
     }
   }