+++ ui-components/src/page_components/listing/listing_sidebar/events/EventDateSection.tsx
@@ -3,6 +3,11 @@ import { ListingEvent } from "@bloom-housing/backend-core/types"
 import dayjs from "dayjs"
 
 const EventDateSection = (props: { event: ListingEvent }) => {
+  const getRangeString = () => {
+    const startTime = dayjs(props.event.startTime).format("hh:mma")
+    const endTime = dayjs(props.event.endTime).format("hh:mma")
+    return startTime === endTime ? startTime : `${startTime} - ${endTime}`
+  }
   return (
     <>
       {props.event.startTime && (
@@ -10,11 +15,7 @@ const EventDateSection = (props: { event: ListingEvent }) => {
           <span className="inline-block text-tiny uppercase">
             {dayjs(props.event.startTime).format("MMMM D, YYYY")}
           </span>
-          <span className="inline-block text-sm font-bold">
-            {dayjs(props.event.startTime).format("hh:mma") +
-              "-" +
-              dayjs(props.event.endTime).format("hh:mma")}
-          </span>
+          <span className="inline-block text-sm font-bold">{getRangeString()}</span>
         </p>
       )}
     </>